chore: relocate API V1 router file
